core: Introduce ShortFormat trait and impls
